test(reservation): split reservation ITs by capability
